Configured docker scripts to load esphome container and setup development & testing environment.
Configured platformio options to run tests in 'native' and
'esp32' environments.

Both environments' tests are building and running successfully.

Note that DOT_PLATFORMIO env var should be set to point to a
.platformio directory on the host that will be mapped into
this project's root directory. This is where a gig or more
of platformio libraries will be downloaded to. If this env
var is left blank, .platformio/ will be created in this project's
directory on the docker host.